third_party/blink/web_tests/flag-specific/enable-features=PendingBeaconAPI/pending_beacon/pending_beacon-basic-expected.txt

CONSOLE WARNING: The provided value 'FOO' is not a valid enum value of type BeaconMethod.
This is a testharness.js-based test.
PASS Basic configuration of a pending beacon.
PASS Verify that beacons default to POST if no options are provided.
Harness: the test ran to completion.
